requirements.txt
 pyduinocli
 pyserial
 pyqt5
 numpy
 pyqtgraph
